rename filters, update dropdown menus and add a little blur when popover is opened
