+++ apps/web/src/components/web/ui/dropdown-menu.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
-import { Check, Circle } from "lucide-react"
+import { Check, ChevronRightIcon, Circle } from "lucide-react"
 import type { ComponentProps } from "react"
 import { cx } from "~/utils/cva"
 
@@ -121,6 +121,46 @@ const DropdownMenuShortcut = ({ className, ...props }: ComponentProps<"span">) =
   return <span className={cx("ml-auto text-xs tracking-widest opacity-60", className)} {...props} />
 }
 
+const DropdownMenuSub = DropdownMenuPrimitive.Sub
+
+const DropdownMenuSubTrigger = ({
+  className,
+  inset,
+  children,
+  ...props
+}: ComponentProps<typeof DropdownMenuPrimitive.SubTrigger> & {
+  inset?: boolean
+}) => {
+  return (
+    <DropdownMenuPrimitive.SubTrigger
+      className={cx(
+        "flex select-none items-center gap-2 rounded-sm px-2 py-1.5 text-sm outline-none focus:bg-card-dark data-[state=open]:bg-card-dark",
+        inset && "pl-8",
+        className,
+      )}
+      {...props}
+    >
+      {children}
+      <ChevronRightIcon className="ml-auto size-4" />
+    </DropdownMenuPrimitive.SubTrigger>
+  )
+}
+
+const DropdownMenuSubContent = ({
+  className,
+  ...props
+}: ComponentProps<typeof DropdownMenuPrimitive.SubContent>) => {
+  return (
+    <DropdownMenuPrimitive.SubContent
+      className={cx(
+        "z-50 min-w-[8rem] overflow-hidden rounded-md border bg-background p-1 text-muted data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2",
+        className,
+      )}
+      {...props}
+    />
+  )
+}
+
 export {
   DropdownMenu,
   DropdownMenuTrigger,
@@ -134,4 +174,7 @@ export {
   DropdownMenuGroup,
   DropdownMenuPortal,
   DropdownMenuRadioGroup,
+  DropdownMenuSub,
+  DropdownMenuSubTrigger,
+  DropdownMenuSubContent,
 }
